requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
colorama==0.4.4
comtypes==1.1.7
importlib-metadata==2.0.0
iniconfig==1.1.1
IPTCInfo3==2.1.4
more-itertools==8.7.0
olefile==0.46
packaging==20.9
Pillow==8.1.0
pip==21.0.1
pluggy==0.12.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pywin32==300
pywinauto==0.6.8
setuptools==49.6.0.post20210108
six==1.15.0
toml==0.10.1
wheel==0.36.2
wincertstore==0.2
zipp==3.4.0